// ui/mill-grinder-ui/src/api/mill/api.ts
export interface Chat {
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'name'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Chat
     */
    'isFavorite'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'created'?: string;
}
/**
 * 
 * @export
 * @interface ChatMessage
 */
export interface ChatMessage {
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'message'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'role'?: ChatMessageRoleEnum;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof ChatMessage
     */
    'content'?: { [key: string]: any; };
}

export const ChatMessageRoleEnum = {
    User: 'USER',
    Chat: 'CHAT',
    System: 'SYSTEM'
} as const;

export type ChatMessageRoleEnum = typeof ChatMessageRoleEnum[keyof typeof ChatMessageRoleEnum];